Improved the MonoTODO messages, as these will be developer visible
2006-11-16  Miguel de Icaza  <miguel@novell.com>

	* Hashtable.cs: Serialize EqualityComparer.


svn path=/trunk/mcs/; revision=67996

+++ mcs/class/corlib/System.Security.Principal/WindowsIdentity.cs
@@ -278,7 +278,7 @@ namespace System.Security.Principal {
 #if !NET_1_0
 		void ISerializable.GetObjectData (SerializationInfo info, StreamingContext context) 
 		{
-			info.AddValue ("m_userToken", _token);
+			info.AddValue ("m_userToken", (long) _token);
 			// can be null when not resolved
 			info.AddValue ("m_name", _name);
 			info.AddValue ("m_type", _type);
